Табы вынесены в отдельный компонент. При клике на карточку фильма открывается соответствующая страница
