+++ src/components/add-review/add-review.jsx
@@ -6,25 +6,27 @@ import Breadcrumbs from "../breadcrumbs/breadcrumbs";
 import AddReviewForm from "../add-review-form/add-review-form";
 
 const AddReview = (props) => {
-  const {films} = props;
+  const {films, currentFilmId} = props;
+  const currentFilm = films.find((film) => film.id === currentFilmId);
+  const {backgroundImage, name, posterImage} = currentFilm;
 
   return (
     <section className="movie-card movie-card--full">
       <div className="movie-card__header">
         <div className="movie-card__bg">
-          <img src={films[0].backgroundImage} alt={films[0].name} />
+          <img src={backgroundImage} alt={name} />
         </div>
 
         <h1 className="visually-hidden">WTW</h1>
 
         <header className="page-header">
           <PageHeaderLogo />
-          <Breadcrumbs films={films}/>
+          <Breadcrumbs film={currentFilm}/>
           <UserBlock />
         </header>
 
         <div className="movie-card__poster movie-card__poster--small">
-          <img src={films[0].posterImage} alt={films[0].name} width="218" height="327" />
+          <img src={posterImage} alt={name} width="218" height="327" />
         </div>
       </div>
 
@@ -36,6 +38,7 @@ const AddReview = (props) => {
 };
 
 AddReview.propTypes = {
+  currentFilmId: PropTypes.number.isRequired,
   films: PropTypes.arrayOf(PropTypes.shape({
     id: PropTypes.number.isRequired,
     name: PropTypes.string.isRequired,
